+++ backend/db/postgres/modules/management/module.sql
@@ -5,3 +5,4 @@
 \ir ../../schema/26_governance_workflow.sql
 \ir ../../schema/27_impact_analysis.sql
 \ir ../../schema/33_readiness_rule_explanations.sql
+\ir ../../schema/35_canonical_service_fields.sql